make CatalogPartitionSpec a class and remove copy()
