--- flink-table/flink-table-api-java/src/main/java/org/apache/flink/table/catalog/GenericCatalogPartitionSpec.java
@@ -1,74 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.flink.table.catalog;
-
-import java.util.Collections;
-import java.util.HashMap;
-import java.util.Map;
-import java.util.Objects;
-
-import static org.apache.flink.util.Preconditions.checkNotNull;
-
-/**
- * A generic catalog partition spec implementation.
- */
-public class GenericCatalogPartitionSpec implements CatalogPartitionSpec {
-	// <partition key, value>
-	private final Map<String, String> partitionSpec;
-
-	public GenericCatalogPartitionSpec(Map<String, String> partitionSpec) {
-		checkNotNull(partitionSpec, "partitionSpec cannot be null");
-
-		this.partitionSpec = Collections.unmodifiableMap(partitionSpec);
-	}
-
-	@Override
-	public Map<String, String> getPartitionSpec() {
-		return partitionSpec;
-	}
-
-	@Override
-	public CatalogPartitionSpec copy() {
-		return new GenericCatalogPartitionSpec(new HashMap<>(partitionSpec));
-	}
-
-	@Override
-	public boolean equals(Object o) {
-		if (this == o) {
-			return true;
-		}
-
-		if (o == null || getClass() != o.getClass()) {
-			return false;
-		}
-
-		GenericCatalogPartitionSpec that = (GenericCatalogPartitionSpec) o;
-		return partitionSpec.equals(that.partitionSpec);
-	}
-
-	@Override
-	public int hashCode() {
-		return Objects.hash(partitionSpec);
-	}
-
-	@Override
-	public String toString() {
-		return "CatalogPartitionSpec{" + partitionSpec + '}';
-	}
-}
