+++ flink-table/flink-table-common/src/main/java/org/apache/flink/table/catalog/CatalogPartitionSpec.java
@@ -18,26 +18,58 @@
 
 package org.apache.flink.table.catalog;
 
+import java.util.Collections;
 import java.util.Map;
+import java.util.Objects;
+
+import static org.apache.flink.util.Preconditions.checkNotNull;
 
 /**
  * Represents a partition spec object in catalog.
  * Partition columns and values are NOT of strict order, and they need to be re-arranged to the correct order
  * by comparing with a list of strictly ordered partition keys.
  */
-public interface CatalogPartitionSpec {
+public class CatalogPartitionSpec {
+
+	// An unmodifiable map as <partition key, value>
+	private final Map<String, String> partitionSpec;
+
+	public CatalogPartitionSpec(Map<String, String> partitionSpec) {
+		checkNotNull(partitionSpec, "partitionSpec cannot be null");
+
+		this.partitionSpec = Collections.unmodifiableMap(partitionSpec);
+	}
 
 	/**
 	 * Get the partition spec as key-value map.
 	 *
 	 * @return a map of partition spec keys and values
 	 */
-	Map<String, String> getPartitionSpec();
+	public Map<String, String> getPartitionSpec() {
+		return partitionSpec;
+	}
 
-	/**
-	 * Get a deep copy of the CatalogPartitionSpec instance.
-	 *
-	 * @return a copy of CatalogPartitionSpec instance
-	 */
-	CatalogPartitionSpec copy();
+	@Override
+	public boolean equals(Object o) {
+		if (this == o) {
+			return true;
+		}
+
+		if (o == null || getClass() != o.getClass()) {
+			return false;
+		}
+
+		CatalogPartitionSpec that = (CatalogPartitionSpec) o;
+		return partitionSpec.equals(that.partitionSpec);
+	}
+
+	@Override
+	public int hashCode() {
+		return Objects.hash(partitionSpec);
+	}
+
+	@Override
+	public String toString() {
+		return "CatalogPartitionSpec{" + partitionSpec + '}';
+	}
 }
